requirements.txt
dify_plugin>=0.1.0,<0.2.0
markdown~=3.4.1
python-docx~=1.1.2
htmldocx~=0.0.6
aiohttp~=3.10.11
xhtml2pdf~=0.2.17